+++ anoexpress/__init__.py
@@ -1,3 +1,6 @@
 # flake8: noqa
 from . import anoexpress
-from .anoexpress import *
+from .data import *
+from .gsea import *
+from .utils import *
+from .plot import *
